+++ src/app/routes/connector-ui/contract-definition-page/contract-definition-editor-dialog/contract-definition-editor-dialog-form.ts
@@ -1,6 +1,7 @@
 import {Injectable} from '@angular/core';
 import {FormBuilder, FormGroup, Validators} from '@angular/forms';
 import {PolicyDefinitionDto} from '@sovity.de/edc-client';
+import {DataOfferFormValidators} from 'src/app/core/validators/data-offer-form-validators';
 import {UiAssetMapped} from '../../../../core/services/models/ui-asset-mapped';
 import {noWhitespacesOrColonsValidator} from '../../../../core/validators/no-whitespaces-or-colons-validator';
 import {
@@ -22,11 +23,18 @@ export class ContractDefinitionEditorDialogForm {
     return this.group.value;
   }
 
-  constructor(private formBuilder: FormBuilder) {}
+  constructor(
+    private formBuilder: FormBuilder,
+    private validators: DataOfferFormValidators,
+  ) {}
 
   buildFormGroup(): FormGroup<ContractDefinitionEditorDialogFormModel> {
     return this.formBuilder.nonNullable.group({
-      id: ['', [Validators.required, noWhitespacesOrColonsValidator]],
+      id: [
+        '',
+        [Validators.required, noWhitespacesOrColonsValidator],
+        [this.validators.contractDefinitionIdExistsValidator],
+      ],
       accessPolicy: [null as PolicyDefinitionDto | null, Validators.required],
       contractPolicy: [null as PolicyDefinitionDto | null, Validators.required],
       assets: [[] as UiAssetMapped[], Validators.required],